+++ lib/msprime.c
@@ -329,6 +329,13 @@ out:
 }
 
 int
+msp_set_stop_at_local_mrca(msp_t *self, bool stop_at_local_mrca)
+{
+    self->stop_at_local_mrca = stop_at_local_mrca;
+    return 0;
+}
+
+int
 msp_set_discrete_genome(msp_t *self, bool is_discrete)
 {
     self->discrete_genome = is_discrete;
@@ -1030,6 +1037,7 @@ msp_alloc(msp_t *self, tsk_table_collection_t *tables, gsl_rng *rng)
     self->store_migrations = false;
     self->additional_nodes = 0;
     self->coalescing_segments_only = true;
+    self->stop_at_local_mrca = true;
     self->avl_node_block_size = 1024;
     self->node_mapping_block_size = 1024;
     self->segment_block_size = 1024;
@@ -3777,6 +3785,7 @@ msp_merge_two_ancestors(msp_t *self, population_id_t population_id, label_id_t l
     bool defrag_required = false;
     tsk_id_t v;
     double l, r, l_min, r_max;
+    uint32_t min_overlap;
     avl_node_t *node;
     node_mapping_t *nm, search;
     segment_t *x, *y, *z, *alpha, *beta;
@@ -3785,6 +3794,12 @@ msp_merge_two_ancestors(msp_t *self, population_id_t population_id, label_id_t l
     if (new_lineage == NULL) {
         ret = MSP_ERR_NO_MEMORY;
         goto out;
+    }
+
+    if (self->stop_at_local_mrca) {
+        min_overlap = 2;
+    } else {
+        min_overlap = 0;
     }
 
     x = a;
@@ -3862,7 +3877,7 @@ msp_merge_two_ancestors(msp_t *self, population_id_t population_id, label_id_t l
                 node = avl_search(&self->overlap_counts, &search);
                 tsk_bug_assert(node != NULL);
                 nm = (node_mapping_t *) node->item;
-                if (nm->value == 2) {
+                if (nm->value == min_overlap) {
                     nm->value = 0;
                     node = node->next;
                     tsk_bug_assert(node != NULL);
@@ -3870,7 +3885,7 @@ msp_merge_two_ancestors(msp_t *self, population_id_t population_id, label_id_t l
                     r = nm->position;
                 } else {
                     r = l;
-                    while (nm->value != 2 && r < r_max) {
+                    while (nm->value != min_overlap && r < r_max) {
                         nm->value--;
                         node = node->next;
                         tsk_bug_assert(node != NULL);
@@ -3983,6 +3998,8 @@ msp_merge_ancestors(msp_t *self, avl_tree_t *Q, population_id_t population_id,
     bool defrag_required = false;
     uint32_t j, h;
     double l, r, r_max, next_l, l_min;
+    uint32_t min_overlap;
+
     avl_node_t *node;
     node_mapping_t *nm, search;
     segment_t *x, *z, *alpha;
@@ -4072,7 +4089,12 @@ msp_merge_ancestors(msp_t *self, avl_tree_t *Q, population_id_t population_id,
             node = avl_search(&self->overlap_counts, &search);
             tsk_bug_assert(node != NULL);
             nm = (node_mapping_t *) node->item;
-            if (nm->value == h) {
+            if (self->stop_at_local_mrca) {
+                min_overlap = h;
+            } else {
+                min_overlap = 0;
+            }
+            if (nm->value == min_overlap) {
                 nm->value = 0;
                 node = node->next;
                 tsk_bug_assert(node != NULL);
@@ -4080,8 +4102,8 @@ msp_merge_ancestors(msp_t *self, avl_tree_t *Q, population_id_t population_id,
                 r = nm->position;
             } else {
                 r = l;
-                while (nm->value != h && r < r_max) {
-                    nm->value -= h - 1;
+                while (nm->value != min_overlap && r < r_max) {
+                    nm->value -= min_overlap - 1;
                     node = node->next;
                     tsk_bug_assert(node != NULL);
                     nm = (node_mapping_t *) node->item;
@@ -5189,9 +5211,19 @@ msp_run_coalescent(msp_t *self, double max_time, unsigned long max_events)
         t_wait = GSL_MIN(mig_t_wait,
             GSL_MIN(gc_t_wait, GSL_MIN(gc_left_t_wait, GSL_MIN(re_t_wait, ca_t_wait))));
 
-        if (fixed_event_time == DBL_MAX && t_wait == DBL_MAX) {
+        if (fixed_event_time == DBL_MAX && t_wait == DBL_MAX
+            && self->stop_at_local_mrca) { // TODO add condition
             ret = MSP_ERR_INFINITE_WAITING_TIME;
             goto out;
+        } else if ((fixed_event_time == DBL_MAX && t_wait == DBL_MAX)
+                   && !self->stop_at_local_mrca) {
+
+            if (max_time >= DBL_MAX) {
+                max_time = self->time;
+            }
+            self->time = max_time;
+            ret = MSP_EXIT_COALESCENCE;
+            break;
         }
         random_event_time = self->time + t_wait;
 
@@ -5442,11 +5474,21 @@ msp_dtwf_generation(msp_t *self)
     label_id_t label = 0;
     tsk_id_t parent_nodes[MSP_MAX_PED_PLOIDY];
 
+    bool coalescence_occurred = false;
+    bool recombination_occurred = false;
+    int n_coal;
+    int n_recomb;
+
     for (i = 0; i < 2; i++) {
         avl_init_tree(&Q[i], cmp_segment_queue, NULL);
     }
 
     for (j = 0; j < self->num_populations; j++) {
+
+        coalescence_occurred = false;
+        recombination_occurred = false;
+        n_coal = 0;
+        n_recomb = 0;
 
         pop = &self->populations[j];
         if (avl_count(&pop->ancestors[label]) == 0) {
@@ -5483,6 +5525,10 @@ msp_dtwf_generation(msp_t *self)
             s->next = parents[p];
             s->node = a;
             parents[p] = s;
+            n_coal++;
+            if (n_coal > 1) {
+                coalescence_occurred = true;
+            }
         }
 
         // Iterate through offspring of parent k, adding to avl_tree
@@ -5500,6 +5546,10 @@ msp_dtwf_generation(msp_t *self)
                     ret = msp_dtwf_recombine(self, x, &u[0], &u[1], parent_nodes);
                     if (ret != 0) {
                         goto out;
+                    }
+                    n_recomb++;
+                    if (n_recomb > 1) {
+                        recombination_occurred = true;
                     }
                     for (i = 0; i < 2; i++) {
                         if (u[i] != NULL && u[i] != x) {
@@ -5542,6 +5592,9 @@ msp_dtwf_generation(msp_t *self)
 out:
     msp_safe_free(parents);
     msp_safe_free(segment_mem);
+    if (coalescence_occurred == false && recombination_occurred == false && ret == 0) {
+        return MSP_DTWF_DID_NOTHING;
+    }
     return ret;
 }
 
@@ -5618,6 +5671,7 @@ msp_run_dtwf(msp_t *self, double max_time, unsigned long max_events)
     avl_tree_t *nodes;
     /* Only support a single structured coalescent label at the moment */
     label_id_t label = 0;
+    bool migration_occurred;
 
     tsk_bug_assert(self->recomb_mass_index == NULL);
     tsk_bug_assert(self->gc_mass_index == NULL);
@@ -5649,6 +5703,8 @@ msp_run_dtwf(msp_t *self, double max_time, unsigned long max_events)
             goto out;
         }
         self->time++;
+
+        migration_occurred = false;
 
         /* Following SLiM, we perform migrations prior to selecting
          * parents for the current generation */
@@ -5699,6 +5755,7 @@ msp_run_dtwf(msp_t *self, double max_time, unsigned long max_events)
                 mig_dest_pop = (population_id_t) k;
 
                 for (i = 0; i < n[k]; i++) {
+                    migration_occurred = true;
                     ret = msp_store_simultaneous_migration_events(
                         self, nodes, mig_source_pop, label);
                     if (ret != 0) {
@@ -5715,6 +5772,7 @@ msp_run_dtwf(msp_t *self, double max_time, unsigned long max_events)
                 nodes = &node_trees[j * self->num_populations + k];
                 mig_source_pop = (population_id_t) j;
                 mig_dest_pop = (population_id_t) k;
+                migration_occurred = true;
                 ret = msp_simultaneous_migration_event(
                     self, nodes, mig_source_pop, mig_dest_pop);
                 if (ret != 0) {
@@ -5737,13 +5795,15 @@ msp_run_dtwf(msp_t *self, double max_time, unsigned long max_events)
                 goto out;
             }
             ret = msp_apply_demographic_events(self, self->next_demographic_event->time);
+            migration_occurred
+                = true; // Demographic events can change population structure
             if (ret != 0) {
                 goto out;
             }
         }
         self->time = cur_time;
         ret = msp_dtwf_generation(self);
-        if (ret != 0) {
+        if (ret != 0 && ret != MSP_DTWF_DID_NOTHING) {
             goto out;
         }
 
@@ -5758,11 +5818,26 @@ msp_run_dtwf(msp_t *self, double max_time, unsigned long max_events)
             }
             self->next_sampling_event++;
         }
+
+        if (ret == MSP_DTWF_DID_NOTHING && migration_occurred == false
+            && self->stop_at_local_mrca == false && msp_get_num_ancestors(self) == 1) {
+            /* No events occurred: we are done */
+            if (max_time >= DBL_MAX) {
+                max_time = self->time;
+            }
+            self->time = max_time;
+            ret = MSP_EXIT_COALESCENCE;
+            break;
+        }
     }
 out:
     msp_safe_free(node_trees);
     msp_safe_free(n);
     msp_safe_free(mig_tmp);
+    if (ret == MSP_DTWF_DID_NOTHING) {
+        // No events occurred: it is still ok
+        return 0;
+    }
     return ret;
 }